Fix --baseurl handling
If an override URL is specified, and it does not end with a trailing slash,
add a trailing slash.

The present logic assumes the trailing slash is there, but doesn't mention
this anywhere, and you end up with corrupted URLs in the metadata if it's
missing the and host and prefix get smushed together.